+++ src/main/java/io/enderdev/endermodpacktweaks/features/crashinfo/InfoBuilder.java
@@ -1,24 +1,12 @@
 package io.enderdev.endermodpacktweaks.features.crashinfo;
 
-import com.google.gson.Gson;
-import com.google.gson.GsonBuilder;
-import com.google.gson.JsonIOException;
-import com.google.gson.JsonSyntaxException;
 import io.enderdev.endermodpacktweaks.EMTConfig;
-import io.enderdev.endermodpacktweaks.EnderModpackTweaks;
+import io.enderdev.endermodpacktweaks.utils.EmtModpackInfo;
 import net.minecraftforge.fml.common.ICrashCallable;
-
-import java.io.File;
-import java.io.FileReader;
-import java.io.IOException;
 
 public class InfoBuilder implements ICrashCallable {
     private final String name;
     private final StringBuilder builder = new StringBuilder();
-
-    private String modpackName = EMTConfig.MODPACK.CRASH_INFO.modpackName;
-    private String modpackVersion = EMTConfig.MODPACK.CRASH_INFO.modpackVersion;
-    private String modpackAuthor = EMTConfig.MODPACK.CRASH_INFO.modpackAuthor;
 
     public InfoBuilder(String name) {
         this.name = name;
@@ -26,38 +14,15 @@ public class InfoBuilder implements ICrashCallable {
 
     @Override
     public String call() throws Exception {
-        boolean sourcedFromManifest = false;
-        if (EMTConfig.MODPACK.CRASH_INFO.readFromManifest) {
-            Gson gson = new GsonBuilder().create();
-            File f = new File("manifest.json");
-            if (!f.exists()) {
-                EnderModpackTweaks.LOGGER.warn("manifest.json file not found, trying minecraftinstance.json");
-                f = new File("minecraftinstance.json");
-            }
-            if (f.exists()) {
-                try (FileReader reader = new FileReader(f)) {
-                    ManifestInfo manifestInfo = gson.fromJson(reader, ManifestInfo.class);
-                    if (manifestInfo != null) {
-                        modpackName = manifestInfo.name;
-                        modpackVersion = manifestInfo.version;
-                        modpackAuthor = manifestInfo.author;
-                        sourcedFromManifest = true;
-                    }
-                } catch (JsonSyntaxException | JsonIOException | IOException e) {
-                    EnderModpackTweaks.LOGGER.error("Failed to read manifest file", e);
-                }
-            } else {
-                EnderModpackTweaks.LOGGER.warn("Manifest file not found");
-            }
-        }
+        EmtModpackInfo modpackInfo = new EmtModpackInfo(EMTConfig.MODPACK.CRASH_INFO.readFromManifest);
 
         builder.append("---\n");
-        builder.append("Modpack Name: '").append(modpackName).append("'\n");
-        builder.append("Modpack Version: '").append(modpackVersion).append("'\n");
-        builder.append("Modpack Author: '").append(modpackAuthor).append("'\n");
+        builder.append("Modpack Name: '").append(modpackInfo.getName()).append("'\n");
+        builder.append("Modpack Version: '").append(modpackInfo.getVersion()).append("'\n");
+        builder.append("Modpack Author: '").append(modpackInfo.getAuthor()).append("'\n");
         builder.append("---\n");
 
-        if (!sourcedFromManifest && EMTConfig.MODPACK.CRASH_INFO.readFromManifest) {
+        if (modpackInfo.error()) {
             builder.append("!!! Failed to read manifest file, using config values instead. !!!\n");
         }
 
@@ -67,11 +32,5 @@ public class InfoBuilder implements ICrashCallable {
     @Override
     public String getLabel() {
         return name;
-    }
-
-    private static class ManifestInfo {
-        String name = null;
-        String version = null;
-        String author = null;
     }
 }